Refactor code contained in 'utils.kt'
Move the token creation function to its own factory class.
